amend cr comments - move uplink controllers location
