+++ src/Tickets/Seating/Controller.php
@@ -12,7 +12,6 @@ namespace TEC\Tickets\Seating;
 
 use TEC\Common\Contracts\Provider\Controller as Controller_Contract;
 use TEC\Tickets\Seating\Frontend\Session;
-use TEC\Tickets\Libraries\Uplink\SLR_Controller as SLR_Uplink_Controller;
 
 /**
  * Class Controller
@@ -119,7 +118,7 @@ class Controller extends Controller_Contract {
 		$this->container->register( Tables::class );
 		$this->container->register( Assets::class );
 
-		$this->container->register( SLR_Uplink_Controller::class );
+		$this->container->register( Libraries\Uplink::class );
 
 		// Manage Order and Attendee data.
 		$this->container->register( Orders\Controller::class );
